+++ manul/src/testing/run_sync.rs
@@ -1,4 +1,4 @@
-use alloc::{collections::BTreeMap, vec::Vec};
+use alloc::{collections::BTreeMap, format, string::String, vec::Vec};
 
 use rand::Rng;
 use rand_core::CryptoRngCore;
@@ -8,8 +8,8 @@ use tracing::debug;
 use crate::{
     protocol::{EntryPoint, Protocol},
     session::{
-        CanFinalize, LocalError, Message, RoundAccumulator, RoundOutcome, Session, SessionId, SessionParameters,
-        SessionReport,
+        CanFinalize, LocalError, Message, RoundAccumulator, RoundOutcome, Session, SessionId, SessionOutcome,
+        SessionParameters, SessionReport,
     },
 };
 
@@ -92,7 +92,7 @@ where
 pub fn run_sync<EP, SP>(
     rng: &mut impl CryptoRngCore,
     entry_points: Vec<(SP::Signer, EP)>,
-) -> Result<BTreeMap<SP::Verifier, SessionReport<EP::Protocol, SP>>, LocalError>
+) -> Result<ExecutionResult<EP::Protocol, SP>, LocalError>
 where
     EP: EntryPoint<SP::Verifier>,
     SP: SessionParameters,
@@ -155,14 +155,47 @@ where
         }
     }
 
-    let mut outcomes = BTreeMap::new();
+    let mut reports = BTreeMap::new();
     for (verifier, state) in states {
-        let outcome = match state {
+        let report = match state {
             State::InProgress { session, accum } => session.terminate(accum)?,
             State::Finished(report) => report,
         };
-        outcomes.insert(verifier, outcome);
+        reports.insert(verifier, report);
     }
 
-    Ok(outcomes)
+    Ok(ExecutionResult { reports })
+}
+
+/// The result of a protocol execution on a set of nodes.
+#[derive(Debug)]
+pub struct ExecutionResult<P: Protocol, SP: SessionParameters> {
+    pub reports: BTreeMap<SP::Verifier, SessionReport<P, SP>>,
+}
+impl<P, SP> ExecutionResult<P, SP>
+where
+    P: Protocol,
+    SP: SessionParameters,
+{
+    pub fn results(self) -> Result<BTreeMap<SP::Verifier, P::Result>, String> {
+        let mut report_strings = Vec::new();
+        let mut results = BTreeMap::new();
+
+        for (id, report) in self.reports.into_iter() {
+            match report.outcome {
+                SessionOutcome::Result(result) => {
+                    results.insert(id, result);
+                }
+                _ => {
+                    report_strings.push(format!("Id: {:?}\n{}", id, report.brief()));
+                }
+            }
+        }
+
+        if report_strings.is_empty() {
+            Ok(results)
+        } else {
+            Err(report_strings.join("\n"))
+        }
+    }
 }