+++ manul/src/tests/partial_echo.rs
@@ -217,11 +217,10 @@ fn partial_echo() {
     let my_subscriber = tracing_subscriber::fmt()
         .with_env_filter(EnvFilter::from_default_env())
         .finish();
-    let reports = tracing::subscriber::with_default(my_subscriber, || {
-        run_sync::<_, TestSessionParams<BinaryFormat>>(&mut OsRng, entry_points).unwrap()
+    let _results = tracing::subscriber::with_default(my_subscriber, || {
+        run_sync::<_, TestSessionParams<BinaryFormat>>(&mut OsRng, entry_points)
+            .unwrap()
+            .results()
+            .unwrap()
     });
-
-    for (_id, report) in reports {
-        assert!(report.result().is_some());
-    }
 }